Fix output formatting (#14)
* fix output formatting, move bot to a subdirectory

as more files end up in the go working directory, there've been issues
with golang tooling running out of memory since it's trying to evaluate
the entire tree. this moves the bot into its own directory to avoid
that.

this also fixes a bug i introduced where formatting was no longer being
correctly applied since we moved from uint64 -> big.Int, handling this
now with a custom formatter. adds tests for that formatter. more are
needed but this is fine for now.

* run bot on PR, but don't commit